itest: Move dcrlnd-only test to own file
The upstream testRevokedCloseRetributionRemoteHodl has actually two
redemption paths that may activated dependent on timing. During the
initial porting effort, this test was split in dcrlnd into two, where
each test verifies one of the redemption paths independently.

This commit moves the dcrlnd-only test to its own file, to ease porting
efforts of upstream code.
